requirements.txt
absl-py==1.2.0
anyio==3.6.1
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
colorama==0.4.5
dlib==19.24.0
fastapi==0.82.0
fastapi-utils==0.2.1
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.48.1
h11==0.13.0
h5py==3.7.0
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
imutils==0.5.4
Jinja2==3.1.2
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydentic==0.0.1.dev3
PyMySQL==1.0.2
pyparsing==3.0.9
pytesseract==0.3.10
python-dotenv==0.21.0
python-multipart==0.0.5
python-stdnum==1.17
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
sentiment-analysis-spanish==0.0.25
six==1.16.0
sklearn==0.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.19.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
